extract trim function to utils
